# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/investigations:
    post:
      operationId: createPersistentInvestigation
      tags: [cashnet]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/InvestigationInput" } } } }
      responses: { "201": { description: Created; collection requires approval and authorization, content: { application/json: { schema: { $ref: "#/components/schemas/PersistentInvestigation" } } } } }
  /v1/investigations/wallet:
    post:
      operationId: createWalletSubjectInvestigation
      tags: [cashnet]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WalletInvestigationInput" } } } }
      responses: { "201": { description: Created; collection requires approval and authorization, content: { application/json: { schema: { $ref: "#/components/schemas/WalletInvestigationResult" } } } } }
  /v1/investigations/{id}:
    get:
      operationId: getPersistentInvestigation
      tags: [cashnet]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Investigation, content: { application/json: { schema: { $ref: "#/components/schemas/PersistentInvestigation" } } } } }
    patch:
      operationId: transitionPersistentInvestigation
      tags: [cashnet]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/InvestigationTransitionInput" } } } }
      responses: { "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/PersistentInvestigation" } } } } }
  /v1/investigations/{id}/collect:
    post:
      operationId: collectInvestigationProviderData
      tags: [cashnet]
      summary: Collect authorized live blockchain facts and persist normalized results
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Collection result, content: { application/json: { schema: { $ref: "#/components/schemas/CollectionResult" } } } } }
  /v1/investigations/{id}/graph:
    get:
      operationId: traceInvestigationGraph
      tags: [cashnet]
      summary: Trace stored blockchain relationships with bounded BFS
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: depth, in: query, schema: { type: integer, minimum: 1, maximum: 5, default: 2 } }
        - { name: direction, in: query, schema: { type: string, enum: [OUTGOING, INCOMING, BOTH], default: OUTGOING } }
        - { name: max_neighbors, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: max_nodes, in: query, schema: { type: integer, minimum: 1, maximum: 1000, default: 250 } }
        - { name: max_edges, in: query, schema: { type: integer, minimum: 1, maximum: 2000, default: 500 } }
        - { name: min_amount, in: query, schema: { type: string, pattern: '^\\d+(\\.\\d+)?$' } }
        - { name: max_amount, in: query, schema: { type: string, pattern: '^\\d+(\\.\\d+)?$' } }
        - { name: asset, in: query, schema: { type: string } }
        - { name: start_time, in: query, schema: { type: string, format: date-time } }
        - { name: end_time, in: query, schema: { type: string, format: date-time } }
      responses: { "200": { description: Evidence-backed graph from stored normalized facts, content: { application/json: { schema: { $ref: "#/components/schemas/InvestigationGraph" } } } } }
  /v1/investigations/{id}/address-intelligence/{chain}/{address}:
    get:
      operationId: lookupInvestigationAddressIntelligence
      tags: [cashnet]
      summary: Look up approved, case-scoped address intelligence observations
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: chain, in: path, required: true, schema: { type: string, enum: [BITCOIN, ETHEREUM, TRON] } }
        - { name: address, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Observations and explicit conflicts; never an identity claim, content: { application/json: { schema: { $ref: "#/components/schemas/AddressIntelligenceLookup" } } } } }
  /v1/investigations/{id}/clusters:
    post:
      operationId: analyzeInvestigationBitcoinClusters
      tags: [cashnet]
      summary: Run bounded, explainable Bitcoin cluster inference over stored facts
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: false, content: { application/json: { schema: { $ref: "#/components/schemas/ClusterRunInput" } } } }
      responses: { "200": { description: Review-required inferences, content: { application/json: { schema: { $ref: "#/components/schemas/ClusterRunResult" } } } } }
    get:
      operationId: listInvestigationClusters
      tags: [cashnet]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }]
      responses: { "200": { description: Stored review-required cluster inferences, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ClusterInference" } } } } } }
  /v1/investigations/{id}/vasp-analysis:
    post:
      operationId: analyzeInvestigationVaspCandidates
      tags: [cashnet]
      summary: Run deterministic evidence fusion for service and VASP candidates
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: false, content: { application/json: { schema: { $ref: "#/components/schemas/VaspAnalysisInput" } } } }
      responses: { "200": { description: "Investigative candidates, not person attribution", content: { application/json: { schema: { $ref: "#/components/schemas/VaspAnalysisResult" } } } } }
  /v1/investigations/{id}/vasp-candidates:
    get:
      operationId: listInvestigationVaspCandidates
      tags: [cashnet]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }]
      responses: { "200": { description: Persisted candidate records with traceable evidence, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/VaspCandidate" } } } } } }
...
components:
  schemas:
...
    PersistentInvestigation:
      type: object
      required: [id, caseId, status, investigationDepth, createdAt, updatedAt]
      properties:
        id: { type: string }
        caseId: { type: string }
        status: { type: string, enum: [CREATED, AUTHORIZED, RUNNING, COMPLETED, PARTIAL, FAILED, CANCELLED] }
        chain: { type: [string, "null"] }
        walletAddress: { type: [string, "null"] }
        investigationDepth: { type: number }
        startTime: { type: [string, "null"], format: date-time }
        endTime: { type: [string, "null"], format: date-time }
        createdBy: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    InvestigationInput:
      type: object
      required: [caseId]
      properties:
        caseId: { type: string }
        chain: { type: string }
        walletAddress: { type: string }
        investigationDepth: { type: number, minimum: 1, maximum: 10 }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
    InvestigationTransitionInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [AUTHORIZED, RUNNING, COMPLETED, PARTIAL, FAILED, CANCELLED] }
    WalletInvestigationInput:
      allOf:
        - $ref: "#/components/schemas/InvestigationInput"
        - type: object
          required: [chain, walletAddress]
          properties:
            label: { type: string, enum: [REPORTED, SUSPECT, SUBJECT, OBSERVED, UNKNOWN] }
    WalletSubject:
      type: object
      required: [id, caseId, investigationId, chain, walletAddress, label, createdAt]
      properties:
        id: { type: string }
        caseId: { type: string }
        investigationId: { type: string }
        chain: { type: string }
        walletAddress: { type: string }
        label: { type: string }
        createdAt: { type: string, format: date-time }
    WalletInvestigationResult:
      type: object
      required: [investigation, walletSubject]
      properties:
        investigation: { $ref: "#/components/schemas/PersistentInvestigation" }
        walletSubject: { $ref: "#/components/schemas/WalletSubject" }
...
    CollectionResult:
      type: object
      required: [investigationId, status, provider, transactionCount, tokenTransferCount]
      properties:
        investigationId: { type: string }
        status: { type: string }
        provider: { type: string }
        transactionCount: { type: number }
        tokenTransferCount: { type: number }
...
    GraphEvidence:
      type: object
      required: [transactionHash, method, derivationSourceType]
      properties:
        transactionHash: { type: string }
        provider: { type: [string, "null"] }
        sourceReference: { type: [string, "null"] }
        rawReference: { type: [string, "null"] }
        retrievedAt: { type: [string, "null"], format: date-time }
        method: { type: string }
        derivationSourceType: { type: string, enum: [API, INFERENCE] }
    InvestigationGraphNode:
      type: object
      required: [id, chain, address, nodeType]
      properties:
        id: { type: string }
        chain: { type: string }
        address: { type: string }
        nodeType: { type: string, enum: [EOA, ADDRESS, CONTRACT, UNKNOWN] }
        firstSeen: { type: [string, "null"], format: date-time }
        lastSeen: { type: [string, "null"], format: date-time }
    InvestigationGraphEdge:
      type: object
      required: [id, chain, transactionHash, fromAddress, toAddress, relationshipType, asset, amount, evidence]
      properties:
        id: { type: string }
        chain: { type: string }
        transactionHash: { type: string }
        fromAddress: { type: string }
        toAddress: { type: string }
        relationshipType: { type: string, enum: [TRANSFER, TOKEN_TRANSFER, INTERNAL_TRANSFER, CONTRACT_INTERACTION, UTXO_SPEND] }
        asset: { type: string }
        amount: { type: string }
        tokenContract: { type: [string, "null"] }
        timestamp: { type: [string, "null"], format: date-time }
        blockNumber: { type: [string, "null"] }
        status: { type: [string, "null"] }
        evidence: { $ref: "#/components/schemas/GraphEvidence" }
    InvestigationGraphPath:
      type: object
      required: [rank, nodes, edgeIds, hopCount, evidenceComplete]
      properties:
        rank: { type: number, minimum: 1 }
        nodes: { type: array, items: { type: object, required: [chain, address], properties: { chain: { type: string }, address: { type: string } } } }
        edgeIds: { type: array, items: { type: string } }
        hopCount: { type: number, minimum: 0 }
        evidenceComplete: { type: boolean }
    InvestigationGraph:
      type: object
      required: [status, nodes, edges, paths, metadata, limitsApplied, evidenceReferences]
      properties:
        status: { type: string, enum: [OK, INSUFFICIENT_DATA] }
        nodes: { type: array, items: { $ref: "#/components/schemas/InvestigationGraphNode" } }
        edges: { type: array, items: { $ref: "#/components/schemas/InvestigationGraphEdge" } }
        paths: { type: array, items: { $ref: "#/components/schemas/InvestigationGraphPath" } }
        metadata: { type: object, additionalProperties: true }
        limitsApplied: { type: object, additionalProperties: true }
        evidenceReferences: { type: array, items: { $ref: "#/components/schemas/GraphEvidence" } }
    AddressIntelligenceObservation:
      type: object
      required: [id, chain, address, entityType, source, retrievedAt, freshnessStatus, confidence, status]
      properties:
        id: { type: string }
        chain: { type: string }
        address: { type: string }
        label: { type: [string, "null"] }
        entityName: { type: [string, "null"] }
        entityType: { type: string, enum: [EXCHANGE, VASP, CUSTODIAL_SERVICE, DEX, BRIDGE, MIXER, MINING_POOL, DEFI, SCAM, PHISHING, SANCTIONED_ENTITY, OTHER, UNKNOWN] }
        source: { type: string }
        sourceReference: { type: [string, "null"] }
        sourceUrl: { type: [string, "null"] }
        datasetName: { type: [string, "null"] }
        datasetVersion: { type: [string, "null"] }
        license: { type: [string, "null"] }
        retrievedAt: { type: string, format: date-time }
        freshnessStatus: { type: string, enum: [FRESH, STALE, EXPIRED, UNKNOWN] }
        confidence: { type: number, minimum: 0, maximum: 1 }
        status: { type: string, enum: [UNKNOWN, ACTIVE, STALE, CONFLICTING, REVIEW_REQUIRED] }
    AddressIntelligenceLookup:
      type: object
      required: [status, observations, conflicts]
      properties:
        status: { type: string, enum: [SUCCESS, NOT_CONFIGURED, UNAVAILABLE] }
        observations: { type: array, items: { $ref: "#/components/schemas/AddressIntelligenceObservation" } }
        conflicts: { type: array, items: { type: object, additionalProperties: true } }
    ClusterRunInput:
      type: object
      properties: { max_transactions: { type: integer, minimum: 1, maximum: 100, default: 50 } }
    ClusterInference:
      type: object
      required: [id, chain, clusterKey, method, methodVersion, confidenceLevel, numericScore, reviewStatus, evidence, members]
      properties:
        id: { type: string }
        clusterKey: { type: string }
        chain: { type: string, enum: [BITCOIN] }
        method: { type: string }
        methodVersion: { type: string }
        confidenceLevel: { type: string, enum: [UNKNOWN, POSSIBLE, LIKELY] }
        numericScore: { type: number, minimum: 0, maximum: 100 }
        reviewStatus: { type: string, enum: [PENDING_REVIEW, ACCEPTED, REJECTED] }
        ambiguityReason: { type: [string, "null"] }
        evidence: { type: array, items: { type: object, additionalProperties: true } }
        members: { type: array, items: { type: object, additionalProperties: true } }
    ClusterRunResult:
      type: object
      required: [status, analyzedTransactions, inferences, truncated]
      properties:
        status: { type: string, enum: [OK, INSUFFICIENT_DATA] }
        analyzedTransactions: { type: integer }
        inferences: { type: array, items: { $ref: "#/components/schemas/ClusterInference" } }
        truncated: { type: boolean }
    VaspAnalysisInput:
      type: object
      properties:
        max_addresses: { type: integer, minimum: 1, maximum: 250, default: 100 }
        max_candidates: { type: integer, minimum: 1, maximum: 250, default: 50 }
    AttributionEvidence:
      type: object
      required: [category, evidenceType, subjectType, subjectId, polarity, contribution, method, methodVersion]
      properties:
        category: { type: string, enum: [DIRECT_BLOCKCHAIN_FACT, GRAPH_EVIDENCE, ADDRESS_INTELLIGENCE, CLUSTER_INFERENCE, ABUSE_INTELLIGENCE, SOURCE_AGREEMENT, SOURCE_QUALITY] }
        evidenceType: { type: string }
        subjectType: { type: string }
        subjectId: { type: string }
        polarity: { type: string, enum: [SUPPORTING, NEGATIVE, CONTRADICTORY] }
        contribution: { type: number }
        source: { type: [string, "null"] }
        sourceReference: { type: [string, "null"] }
        sourceUrl: { type: [string, "null"] }
        retrievedAt: { type: [string, "null"], format: date-time }
        method: { type: string }
        methodVersion: { type: string }
        rawReference: { type: [string, "null"] }
        details: { type: object, additionalProperties: true }
    VaspCandidate:
      type: object
      required: [id, chain, address, entityType, confidenceLevel, numericScore, status, reason, contradictions, method, methodVersion, evidence]
      properties:
        id: { type: string }
        chain: { type: string }
        address: { type: string }
        entityName: { type: [string, "null"] }
        entityType: { type: string }
        confidenceLevel: { type: string, enum: [UNKNOWN, POSSIBLE, LIKELY, CONFIRMED] }
        numericScore: { type: number, minimum: 0, maximum: 100 }
        status: { type: string, enum: [PENDING_REVIEW, CONFLICTING_EVIDENCE, INSUFFICIENT_EVIDENCE, CONFIRMED_BY_REVIEW] }
        reason: { type: string }
        contradictions: { type: array, items: { type: object, additionalProperties: true } }
        method: { type: string }
        methodVersion: { type: string }
        evidence: { type: array, items: { $ref: "#/components/schemas/AttributionEvidence" } }
    VaspAnalysisResult:
      type: object
      required: [status, candidates, truncated]
      properties:
        status: { type: string, enum: [OK, INSUFFICIENT_EVIDENCE] }
        candidates: { type: array, items: { $ref: "#/components/schemas/VaspCandidate" } }
        truncated: { type: boolean }